+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes are documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## Unreleased
+
+* Fixed: Handle upgrades and loading errors gracefully
+
 ## 1.3.0 - 2017-07-21
 
 ### Added